# perf: geting lazy?, doing less work? generate more (plus some recommendations)