--- a/packages/react-router/src/redirects.ts
+++ b/packages/react-router/src/redirects.ts
@@ -26,9 +26,9 @@ export type Redirect<
 export type ResolvedRedirect<
   TRouter extends AnyRouter = RegisteredRouter,
   TFrom extends RoutePaths<TRouter['routeTree']> = '/',
-  TTo extends string = string,
+  TTo extends string = '',
   TMaskFrom extends RoutePaths<TRouter['routeTree']> = TFrom,
-  TMaskTo extends string = string,
+  TMaskTo extends string = '',
 > = PickAsRequired<
   Redirect<TRouter, TFrom, TTo, TMaskFrom, TMaskTo>,
   'code' | 'statusCode' | 'headers'
@@ -39,9 +39,9 @@ export type ResolvedRedirect<
 export function redirect<
   TRouter extends AnyRouter = RegisteredRouter,
   TFrom extends RoutePaths<TRouter['routeTree']> | string = string,
-  TTo extends string = string,
+  TTo extends string = '',
   TMaskFrom extends RoutePaths<TRouter['routeTree']> | string = TFrom,
-  TMaskTo extends string = string,
+  TMaskTo extends string = '',
 >(
   opts: Redirect<TRouter, TFrom, TTo, TMaskFrom, TMaskTo>,
 ): Redirect<TRouter, TFrom, TTo, TMaskFrom, TMaskTo> {