… docs/framework/react/guide/type-safety.md
… examples/react/basic-file-based/src/routeTree.gen.ts
… examples/react/large-file-based/package.json
… examples/react/large-file-based/src/createRoutes.mjs
… examples/react/large-file-based/src/main.tsx
… examples/react/large-file-based/src/routeTree.gen.ts
… examples/react/large-file-based/src/routes/__root.tsx
… examples/react/large-file-based/src/routes/params/$paramsPlaceholder.tsx
… examples/react/large-file-based/src/routes/params/route.tsx
… examples/react/large-file-based/src/routes/search/route.tsx
… examples/react/large-file-based/src/routes/search/searchPlaceholder.tsx
… packages/react-router/tests/fileRoute.test-d.tsx
… packages/react-router/tests/link.test-d.tsx
… packages/react-router/tests/router.test.tsx
… packages/router-generator/tests/generator/append-and-prepend/routeTree.expected.ts
… packages/router-generator/tests/generator/flat/routeTree.expected.ts
… packages/router-generator/tests/generator/nested-layouts/routeTree.expected.ts
… packages/router-generator/tests/generator/nested-layouts/routes/(folder)/in-folder.tsx
… packages/router-generator/tests/generator/nested/routeTree.expected.ts
… packages/router-generator/tests/generator/single-level/routeTree.expected.ts
… pnpm-lock.yaml