@schiller-manuel 2024-05-06
we should add that `What can blow up your TS Performance` section into the documentation