@chorobin 2024-05-06
> we should add that `What can blow up your TS Performance` section into the documentation

Definitely. Think we need to write a section on it so they can go to it and be able to resolve common problems