@schiller-manuel 2024-05-11
@SeanCassiere in light of https://github.com/TanStack/router/pull/1584, can you please review this?