@SeanCassiere 2024-05-11
Also, we probably want to using the `feat:` or `refactor:` prefix on the commit name. Not sure if we are using conventional commits or angular commits, but if I've not mistaken on one of them using `perf:` triggers a major version bump.