@SeanCassiere 2024-05-11
> @SeanCassiere in light of #1584, can you please review this?

Trying out the changes from my previous PR on this yielded no collisions that I could observe.

That being said, @chorobin once the merge conflict has been resolved, you'd want to rebase and confirm the tests I introduced in that PR still work.