@chorobin 2024-05-12
> Also, we probably want to using the `feat:` or `refactor:` prefix on the commit name. Not sure if we are using conventional commits or angular commits, but if I've not mistaken on one of them using `perf:` triggers a major version bump.

I've never heard of this. In conventional commits I thought you had to explicitly mention `!` or `BREAKING CHANGE` to bump major version