@schiller-manuel 2024-05-12
this can be seen here: https://github.com/TanStack/config/blob/main/src/publish/index.js#L144

however, it only matters that the squashed commit has that prefix, so it can be set when this is merged